Add operator-> method to data iterator wrappers
This adds many, but not all, of the methods of the datum the pointer is
pointing at to the pointer object. This means the "curr" method is less
needed (and  there's a __deref__ method that does the same thing).

According to the SWIG documentation all the datum's methods should be
added, but for some reason many of them aren't. They are mostly the
methods each datum inherits from the abstract base class Metadatum, even
though the base class doesn't implement the methods.