Wrap data classes instead of extending them
This makes it easier to benefit from SWIG type conversions, overloaded
functions, and so on, and should make it easier to do more interesting
things in future. The downside is that as the class is defined in a
module it can't be shared, so has to be redefined in the image module.
"Duck typing" means the user shouldn't really notice that.