Use two classes for data iterators
The base class holds a fixed pointer that can't be dereferenced and is
suitable for storing the result of data.end(). A derived class has a
safe pointer than can be incremented but not as far as data.end(). This
pointer can always be safely dereferenced, so all the previous pointer
checking stuff can be got rid of.