Change gettext initialisation
There's no need to call bindtextdomain in a SWIG generated C module as
Python's locale package provides direct access to it. Also moved
setlocale out of the SWIG generated C module as it should not be called
from within a Python library package. Applications should call it, once,
as now shown in the examples.