Remove use of value_ attribute
This largely undoes commit 5f647dc

The problem is that XX.value_ is a pointer to data in XX, so if XX gets
deleted you can be sure of a segfault. Wrapping the Python value_ in a
class that increfs the parent is too much hassle for too little gain.