Transcode Windows paths to current code page
This is an alternative approach to using the std::wstring versions of
exiv2 methods that are not available in v0.28.x (see
https://github.com/Exiv2/exiv2/issues/2637). It should also allow non
ascii paths to be used when libexiv2 has been compiled without wstring
methods enabled.

The proper approach is to set the Windows code page to utf-8, but this
probably won't work on old Windows, and it's not clear how to do it for
a Python application. Instead we transcode paths from utf-8 to the
current code page, which could fail.