Improve documentation of package version info
The `__version__` and `__version_tuple__` attributes are now documented
as part of the exiv2 package, rather than the exiv2.version module.

The functions in the exiv2.version module are explicitly listed as
Sphinx auto summary was confused by there being a function exiv2.version
as well as a package of the same name.